Stop using reflection to create GrpcServerImpl
This dates to a bygone era where the grpc library and/or stubs were unavailable
during bootstrapping. With lib/remote, among others', reliance on grpc, it's
all but certain this ship has sailed. Removing this indirection to remove room
for error.

The RPCServer.Factory abstraction is no longer necessary and is deleted. I kept
the RPCServer interface in place since it seemed like a good way to prevent
GrpcServerImpl from becoming unweildy, and because it was easier than removing
it.

The reflection _was_ still useful for avoiding a circular dep between runtime
and server. This was only useful for some logging, which I pushed into another
library to fix the glitch.

PiperOrigin-RevId: 331135247