Fix test_interrupt_streamed_output_* flakiness.
This test waits for a Bazel subprocess to reach a point where it can be
interrupted and then sends an interrupt to Bazel, expecting the process to
exit cleanly.

To do this, the test waits for the subprocess to print a message before
entering a long sleep.  The problem is that we do this twice within the same
test.  And this is a problem because the output redirection of the Bazel
invocation happens in the subprocess too, so there is no guarantee that the
redirection has already taken place on the next statement of the script.
As a result, the second Bazel run we do could mistakely pick up the
TEST_log of the previous attempt and assume Bazel had already reached the
interruption point.

To fix this, remove TEST_log before starting Bazel in the background so that
we aren't confused by previous output.

PiperOrigin-RevId: 331773152