Recognize CUDA header file type (.cuh)
This PR mirrors the fix in
https://github.com/bazelbuild/bazel/commit/8804128f52a674b9f8faf3b9534ae156cbad00a1
therefore shouldn't be considered as a standalone PR, but rather a
follow up for the said commit as a bugfix.

.cuh is widely used as suffix for CUDA. It is made more prominent since
"cub" (https://nvlabs.github.io/cub/) recently distributed with CUDA
ToolKit 11, hence, now .cuh headers are in every CUDA ToolKit
distribution.

From that lens, this PR simply fixed the said commit so that newer CUDA
ToolKit can be natively supported in Bazel.

I also believe if this PR made its way in, we can close this issue: https://github.com/bazelbuild/bazel/issues/6578