Clear reusable state in BuildEventArtifactUploaderFactory on shutdown.
An implementation of BuildEventArtifactUploaderFactory is permitted to reuse a
long-lived client to the remote service that receives uploaded artifacts, so long
as the factory respects changes to the relevant flags that configure the client. To
ensure that we gracefully tear down that long-lived client when blaze itself is
shutting down, the factory must expose a shutdown operation that is only called
after all pending uploads have been cancelled.

RELNOTES: None.
PiperOrigin-RevId: 332467417