bazel skyframe: merge WorkspaceASTFunction into WorkspaceFileFunction
The separation between these two Skyframe functions is arbitrary and
problematic. The first used to read, parse, and split the WORKSPACE file
into chunks, and the would resolve and execute a single chunk.
However, resolving is a mutation, which violates Skyframe invariants.

Now, we redo the read/parse/split operation for each chunk.
This is perhaps slightly less efficient, but it creates many
opportunities for later simplification.

For ease of review, this change contains only the minimal edits
required for merging. Later changes will clean things up.

(It is alarming that there are so many copies of the logic
to assemble the list of Skyframe functions known to Blaze.
These should be internal details of each package.)

PiperOrigin-RevId: 332508078