HttpCacheClient: make upload async
This matches the behavior for download, and improves performance for actions
with a lot of outputs.

The code was already using futures in the API and in the implementation,
but the glue code blocked on the internal future and returned an immediate
future to the caller, rather than doing it async.

Fixes #6091.

Change-Id: I3151aa96b879323e0000d3209f6b9bc8be0066d4

Closes #12115.

Change-Id: I72032f7b3e88de4d2e2b5abd2d89380f7fcc31e5
PiperOrigin-RevId: 332796965