Remove an inaccurate comment and simplify a test
Noticed in https://github.com/bazelbuild/bazel/commit/3490011e9c50ea3067d08032e22908d0fa8552f7 (after that was submitted) that a comment saying "in order to depend on the analysistest, we need to return an executable" is a bit inaccurate. An executable needed to be returned because parent_test is a test rule. A rule that depends on a testonly rule must be testonly, and test rules are testonly by default. Just marking the target as testonly makes for a simpler unit-test, since that rule no longer needs an action to generate an executable and providers to return it.

RELNOTES: None.
PiperOrigin-RevId: 332894885