Equip LabelCause with DetailedExitCode, relay detailed input failures in action execution
LabelCauses now specify a DetailedExitCode (like ActionFailed Causes).
They're used during action execution to specify the cause of a missing
input that fails the execution.

MissingFileArtifactValue relays the DetailedExitCode describing why the
file was missing. It no longer contains an exception. The exception was
never used as an exception anyway (just as a bearer of a string).

To be compatible with the required DetailedExitCode in LabelCauses,
ToolchainExceptions now also specify detailed failures.

RELNOTES: None.
PiperOrigin-RevId: 332932518