Stop throwing IncompatibleMergeException merging mismatched branch coverage.
This is a quick hack to unblock internal systems.

C++ header files are able to produce branch coverage information that
differs between compiled object files. Since we are required to merge
these "somehow" and not crash, we just drop one half of the merge in the
conflicting case.

This will need to be improved in the future; the final coverage report
will depend on the order of inputs and may not be preserved across runs.

Improves #12142 by not crashing at least.

PiperOrigin-RevId: 333049821