docgen: move and simplify StarlarkConstructor annotation
Previously, it was in the Starlark interpreter, but it is only used by docgen.

Both its fields have been eliminated because they provide redundant information.
I would have made this change a long time ago but for the fact that
the the information was not entirely redundant. I now believe the only two
places where the provided information differs from trivially computed values
were both mistakes. They are:

1) XcodeConfigInfo, which is now called XcodeVersionConfig.
   The former string appears only in the docs, but never in the codebase.
   See: https://docs.bazel.build/versions/3.4.0/skylark/lib/XcodeVersionConfig.html#XcodeConfigInfo

2) apple_common.AppleStaticLibrary, which is now called AppleStaticLibrary.
   The former name is not an identifier, so it leads to nonsensical documentation.
   See: https://docs.bazel.build/versions/3.5.0/skylark/lib/AppleStaticLibrary.html#members

PiperOrigin-RevId: 333103884