starlark: delete redundant StarlarkDeprecated annotation
Its entire effect was to add the string "(Deprecated)"
to the skydoc output. This is redundant with the doc strings
of all such annotated methods, and uninformative:
Why is it deprecated? Will it be deleted? What should one
use instead?

Also, we already have one complete mechanism for experimental
or deprecated functions and parameters: FlagControlledValue.

PiperOrigin-RevId: 333109275