Java: depend on lcov merger as host
The lcov merger tool used in coverage by Bazel was being depended on in the target
configuration from the Java test rules instead of the host configuration. This
meant that the Java toolchain used in the target configuration would instrument
the lcov merger tool itself with Jacoco.

This was discovered while fixing #4685 because the instrumented lcov merger was
trying to generate a report in the execroot. Another effect of this bug
unnoticed until now was that we always generated an additional empty lcov file
during local coverage execution.

RELNOTES:none
PiperOrigin-RevId: 333471532