Add Support for LLVM's Propeller (basic block sections) Optimization Flags.
Basic Block Sections was recently introduced in LLVM. This supports option
-fbasic-block-sections=list=<file>  which allows basic blocks that are
specified to be placed in unique sections.  This is part of the Propeller
optimization project and is used to perform post link code layout
optimizations :[]

This change adds "propeller_optimize" rule to bazel that allows specifying two
files that would be used to create basic block sections.  The first is called
cc_profile which will be passed to all compile actions.  The second is called
ld_profile which will be passed to the link action.  Both files are expected to
be ".txt".

The cc_profile artifact will be used as input to the
-fbasic-block-sections=list option and the ld_profile artifact will be used as
input to the --symbol-ordering-file link option.  The purpose of this change is
to make sure that these two files are available at the time of the compile/link
action.
PiperOrigin-RevId: 334381480