Remove IncompatibleMergeException from coverage generator
After changing the handling of mismatched branches in coverage reports
to no longer raise an error, there is nothing that throws this
exception.

Even if the handling of branches were to be changed, it has been made
clear that "inconsistent" reports are something that must be handled
without an error being thrown.

PiperOrigin-RevId: 334391902