bazel starlark: merge lib.syntax.util.EvaluationTestCase into BazelETC
//src/test/.../lib/syntax is now deleted.

The BETC diff is messy, even though the file was created by a move.
The only real changes are to the new{Module,Thread}Hook functions and
to the class declaration and its comments.

This change is a step towards a future in which tests do not (and cannot)
concern themselves with subtle internal details such as Starlark thread
creation.

PiperOrigin-RevId: 334417122