Automated rollback of commit 006b4972b433bda280e6982d297da6071b84b88d.
*** Reason for rollback ***

Breaks a number of targets in the nightly.

*** Original change description ***

Make Bazel not raise an error when a symlink points to its own ancestor.

This is only an error when the symlink is to be traversed recursively. E.g. `a/a/a/b` is perfectly acceptable if `a/a` is a symlink to `.`, except if we want to recursively traverse into it during a glob or recursive target pattern expansion.

RELNOTES: Blaze now allows symbolic links that point to their own ancestor unless they are traversed recursively by e.g. a //... recursive target pattern or a recursive glob.
PiperOrigin-RevId: 334579512