Improve toolchain resolution debug
When multiple toolchain types are resolved, skyframe functions run in parallel and the debug output is often hard to follow. Adding type of toolchain to each message makes it possible to analyse the results better.

Also after a execution transition, target changes. Added target label to the message, to be able to tell for which target toolchain was selected.

Removing initial messages "Looking for toolchain of type...", "Considering toolchain...". With additional fields displayed they become obsolete.

Condensing 'rejection explanation', i.e. which values are mismatched, into a single line.

#7713, b/126375095: there is still some duplications in ToolchainResolutionFunction:125, which I didn't remove because some tests depend on it (when specifying extra_toolchains)

b/117279695 is fixed with this

Closes #12196.

PiperOrigin-RevId: 334613787