File names reported as artifact names in BEP are encoded UTF-8.
The build tool has generally inconsistent support for preserving UTF-8 strings
largely due to the absence of UTF-8 support in BUILD files. Most artifact names
in BEP do come from BUILD files where it is assumed UTF-8 identifiers do not
appear. In the cases where artifact names are determined from file names, and the
file name is UTF-8, we may re-encode the generated artifact name as UTF-8 before
adding it to a `named_set_of_files` BEP message.

This issue only leads to incorrect BEP on Unix-like systems due to our treatment
of paths returned by Unix filesystem APIs - we always pass them to Java as
with a Latin-1 encoding. When UTF-8 bytes are in a Latin-1-encoding String and
are passed to a protobuf string field, UTF-8 conversion is applied to UTF-8 bytes.
This does not happen on Windows because we do not produce Java strings with
incorrect encodings.

RELNOTES: None.
PiperOrigin-RevId: 334667088