Delete symlinks in the execroot that are toplevel output directories.
If a directory in the execroot is a toplevel output directory, then it should not be deleted because it was not created as part of symlink forest creation, unless it is a symlink. If the directory is a toplevel output directory and it is a symlink, then this means that it was created as part of a previous build where it was not a toplevel output directory at the time, and should be deleted. If these are not deleted, then certain actions (like actions for creating symlinks to output root inputs) will behave incorrectly (like creating symlinks that point to themselves).

RELNOTES: None.
PiperOrigin-RevId: 334726820