Funnel all JNI load requests through JniLoader.
To avoid symbol collisions in native dependencies, we can safely only load
one shared library to provide our JNI bindings.  Unfortunately, we have
multiple places that attempt to load JNI code, which gives the illusion that
loading more than one library is OK.  To clarify this, centralize all JNI
loads into a single JniLoader entry point.  This change should be a no-op and
does not fix any problems with duplicate symbols yet.

As part of this change, remove the io.bazel.EnableJni JVM system property
that was used to tell Bazel whether to load the JNI code or not... sometimes.
We had many call sites that attempted to load JNI and some checked for this
property and others didn't.  Things worked mostly by chance and were very
fragile, and we can do better by being slightly smarter.

Note that UnixJniLoader and WindowsJniLoader remain after this change but
they are now hidden behind JniLoader.  An upcoming change will further
merge the code of these three, but I'll do it separately for simplicity
reasons.

PiperOrigin-RevId: 335040383