Force /usr/bin/python2.7 on macOS
Because of this issue https://github.com/bazelbuild/bazel/issues/12049
this call to python doesn't respect the PATH users invoke bazel with.
Because the default shell path contains /usr/local/bin, this `python`
could end up being from homebrew. Homebrew's python installations are
easily borked which leads to hashlib being unusable:

https://discuss.bitrise.io/t/broken-python-2-7-hashlib-in-new-xcode-10-3-x-mojave-stack/11401/7

This changes this use case to a path that always exists (at least for
now) on macOS, and only falls back to the bare executable if it doesn't.
Ideally this would utilize the python version from the toolchain
instead.

Closes #12072.

PiperOrigin-RevId: 335109991