Fold the windows/jni subpackage into its parent.
The WindowsJniLoader is now gone, so the windows/jni subpackage only contains
code that has some native calls.  Users of these interfaces shouldn't have to
know if they depend on JNI or not given that this detail has no impact on how
the interfaces are used.  So remove the subpackage for simplicity.

PiperOrigin-RevId: 335128918