Use Java resources to bundle the JNI library.
This gets rid of our custom "nativedeps" code to manually bundle the JNI
library into the Bazel self-extracting executable.  The replacement adds
some complexity to extract it from a Java resource into a file, but the
JniLoader was already complex enough in this regard because Windows, so
this doesn't make things worse.

The benefit of this approach is that the JniLoader library is now
self-contained and does not rely on external actors to do the right thing
regarding placement of the file or configuration of the java.library.path
variable.  In particular, tests can now work without any extra effort on
the caller side.

This change also gets rid of all the target-os-* BUILD logic, which was
added in an attempt to make this whole thing better but which, in
retrospect, made things even more confusing.

You'll notice that I had to change some tests to use a subdirectory of the
temp directory.  That's because this change causes the JniLoader to create
a temporary file in that directory (because Bazel's test runner overrides
the TMPDIR, TMP, and TEMP variables) and would confuse these tests.

PiperOrigin-RevId: 335856030