Support absolute paths for Propeller profiles with propeller_optimize option.
propeller_optimize rule was added to blaze in https://github.com/bazelbuild/bazel/commit/2c7794b5053d849a45e4c73bc93e3004ce8e9e86.  This change allows
absolute paths to be specified for cc_profile and ld_profile.  This is needed
when the profiles are generated in one  snapshot and are used in a
different snapshot.

This is temporary and will be deleted once this feature is ported to Starlark.

PiperOrigin-RevId: 336201673