Don't prematurely add native module to bzl env
Previously, ConfiguredRuleClassProvider#getEnvironment returned an env that contained a binding for "native", mapping to a StarlarkNativeModule instance. This got overwritten in PackageFactory to instead map to a struct containing the same fields (e.g. glob()) plus all the registered rules (e.g. cc_library). Since ASTFileLookupFunction used the getEnvironment() method, it relied on the bogus StarlarkNativeModule object being there for static name validation to work.

As of unknown commit, ASTFileLookupFunction now queries PackageFactory for the final environment, so we are free to no longer add this incomplete dummy object. Except for BzlEnv, StarlarkNativeModule is now only used in conjunction with Starlark.addMethods(), not Starlark.addModule(), so the object itself is almost inaccessible. (Its class is still the home of the user documentation for the special native object fields.)

Verified that ApiExporter produces the same result after this change. (In fact, that target doesn't even depend on this CL.)

Cleanup along the way to #11437 and #11954.

RELNOTES: None
PiperOrigin-RevId: 336211639