Fix action conflicts from exec transition + platform mappings + starlark transition.
Specifically when a two targets under a starlark transition both nullify the effect of the starlark transition (set back to the default configuration pre-starlark transition), one through a platform mapping and one through a platform mapping + an exec transition, make sure they similarly still propagate the effects of the starlark transition (i.e. CoreOptions.transitionDirectoryNameFragment and CoreOptions.affectedByStarlarkTransition)

The dream is to actually have neither propagate the effects of the starlark transition if it's going to nullified. But figuring out when we can do this would be very complex. So for now, just ensure the behavior is the same across the board.

PiperOrigin-RevId: 336362151