Update documentation around newer exit codes
AFAICT we never wound up using "RESEREVED" for what it was "reserved" for.
Since it hasn't been used in recent memory it should be safe to delete - if
someone reuses the code later it won't be a big deal.

The site-specific tool that used 50+ no longer uses those codes, so delete
that comment too.

PiperOrigin-RevId: 336709295