Reduce GC in ActionExecutionFunction caused by removal of MiddlemanArtifacts.
With the targets in b/169092361, turning off middlemen resulted in a 39% increase in GC.time This CL cuts that down to 13%, and also slightly improved the GC time (-3.4%) when Middlemen is enabled.

With NSOS in ActionExecutionFunction, we maintain a map from SkyKey -> Artifacts whose Artifact::key is the SkyKey. There's no need to do that for a big portion of input Artifacts of actions, whose corresponding SkyKeys are just themselves (e.g. SourceArtifacts). This CL removes such Artifacts from the map.

PiperOrigin-RevId: 336882444