Try to avoid conflicts writing param files by doing in-place renames.
Params files are tricky to handle with dynamic scheduling: they are inputs
and we can't take the output lock when we write them out because, otherwise,
we'd defy the purpose of dynamic scheduling.

The --experimental_delay_virtual_input_materialization flag helps fix this
conflict when sandboxing is enabled... but fails to protect us against the
case where we don't use sandboxing, which is the common case with local
persistent workers.

To resolve this, try to write the params files atomically: first write them
to a location that doesn't conflict among strategies and then do an in-place
rename to the final name.  Given that the input is the same and we expect
both strategies to write the same output, this should give us a consistent
view of the file no matter who writes it.  And if it happens that a local
process already started reading from the file before it's replaced by the
remote strategy, no big deal: it's OK to read from unlinked files on Unix.

RELNOTES: None.
PiperOrigin-RevId: 337090508