Fix NullPointer exception caused by attribute.getDefaultValue(null) in StarlarkRuleContext constructor.
The conditions when this bug happened is to have a Starlark aspect over a native aspect, that is using ComputedDefault attributes. This is when StarlarkRuleContext for the native aspect is constructed, calling getDefaultValue(null) in the native aspect.

If only Starlark aspects are involved, getDefaultValue(null) never throws an NPE.

RELNOTES: None.
PiperOrigin-RevId: 337107886