Change action failure description to be action.describe(), rather than its primary output.
Remove longstanding kludges to work around the absence of this information. In some cases, we now print the action's description instead of its owning label. I think that's ok.

Main changes: for C++ compilation and file write actions, we print the usual description instead of their bespoke one which had the label. For errors around output file failures, we print action.describe() instead of the rule label. For errors in include scanning, we print the action, then the subtask ("include scanning"), then the failure.

I've left a little kludge, for old times' sake: because ActionExecutionException's message is used both in SkyframeActionExecutor and in SkyframeBuilder when constructing a BuildFailedException, but there are lots of places that construct an ActionExecutionException, it is left to the ActionExecutionException consumer to prepend action.describe() + " failed: ". We could have ActionExecutionException itself do that in its constructor when it is constructed, but there are subtleties involving its message being null that I didn't want to untangle.

RELNOTES: Error messages emitted when an action fails are reworked to be more informative about the failing action. Some tooling may have to be updated as a result.

Closes #11151

PiperOrigin-RevId: 337305888