bazel profiler: remove TaskStack data structure
It is unnecessary to keep the stack around explicitly when we also can pass the data to the closure.
Previously, the stack was also used to verify that the profiler calls are consistent / properly nested. This became unnecessary with unknown commit which also made `completeTask` private.

Removing `taskStack` also removes the potential for race conditions when stopping the profiler.

PiperOrigin-RevId: 337497234