Implement cquery --output=graph
Thankfully, `query` already has most of the infrastructure necessary to make this
easy.

query implements graph output (in `GraphOutputFormatter`) over a
`Digraph<Target>`, which is a generic graph data structure with `Target` payloads.
All output logic then runs over this data structure. To opt query in, all we have
to do is create an equivalent `Digraph<ConfiguredTarget>`, which is a simple
transformation from the backing graph.

This change creates a new generic class for that common logic:
`GraphOutputWriter`. query's `GraphOutputFormatter` then becomes a simple wrapper
over that, and the new `GraphOutputFormatterCallback` is cquery's equivalent.

A few differences:

 - cquery output is always fully ordered (`--order_output=full`). We could match
   this with query's controllable version, but I don't see a reason to make this
   yet another bit to configured.
 - query output annotates edges with select() conditions. cquery doesn't do this
   because select()s are resolved and removed from the graph after analysis. I
   think we could annotate edges with the *chosen* condition if there was
   demand, but that'd be a followup effort.

Fixes https://github.com/bazelbuild/bazel/issues/10843 (for `cquery`, not `aquery`)

Closes #12248.

PiperOrigin-RevId: 337907070