Fix Bazel clang coverage to produce lcov report
Support for coverage with clang in Bazel was half working because the report produced was in a binary format and it wasn't be utilized by the LCOV merger and therefore didn't show up in the final LCOV report.

This CL uses llvm-cov as was used by Envoy in https://github.com/envoyproxy/envoy/pull/10909 to produce an LCOV report.

To run this the following environment variable and build flags are needed:
BAZEL_USE_LLVM_NATIVE_COVERAGE=1 GCOV={path to llvm_profdata} CC=clang \
    BAZEL_LLVM_COV={path to llvm_cov} bazel coverage --experimental_generate_llvm_lcov
PiperOrigin-RevId: 338254344