Ninja symlink outputs: fix working_dir and output_root path handling.
symlink_outputs in Ninja files are always created relative to the working directory. If the working directory is /path/to/foo, and the output root is /path/to/foo/out, the symlink_outputs values in the Ninja file will be "out/path/to/symlink".

Currently, the code tries to get the relative path of the symlink from the output root ("out/path/to/symlink".relativeTo("out") == "path/to/symlink"), and tries to find that in a set of paths that *do* contain the output root as the prefix, since the set contains path relative to the *working directory*. This causes failures.

Instead, use the path relative to the working directory while checking the symlink outputs set instead, preserving the output root path prefix.

Added a test case where the working dir is the top level directory, and "out" as the toplevel_output_directory and ninja_graph.output_root, emulating the environment where this bug was found.