starlark: improve expectation hygiene in ScriptTest
Prior to this change, ScriptTest inherited a fuzzy syntax for
expectations from its Python-based predecessor: the content
of a ### comment was interpreted either as a substring, or
as a regular expression.

This change causes it to interpret the content of the comment
as a regular expression, always. Some expectations were updated.

PiperOrigin-RevId: 338258986