Remove always failing include path searching
Since this was run as part of the workspace setup, DEVELOPER_DIR and
SDKROOT are not set, which means wrapped_clang, which was being used as
the cc for this, will always fail. Whether or not the invocation fails
isn't taken into account for the result, and the expected output was
never produced, so this always returned nothing.

Closes #11711.

PiperOrigin-RevId: 338260943