Add relative_ast_path_feature
This feature applies for darwin builds to disable absolutizing
swiftmodule paths to link commands. Previously these paths were
absolutized to support debugging. This is not required if users change
the pwd of their debugger so that the relative paths are valid. Since we
don't have a way to force that on users, absolutizing them is still the
default, but users who configure that can enable this feature to improve
hermeticity of their builds.

RELNOTES: Add `relative_ast_path` feature for darwin builds to relativize swiftmodule paths for debugging

Closes #12285.

PiperOrigin-RevId: 338451975