bazel packages: define proto.encode_text(x)
This is the replacement for struct.to_proto, and behaves
the same, but doesn't pollute the struct method namespace.

The implementation has been cleaned up to make the
control simpler to follow and to be less profligate
with allocation. It also rejects dict keys that are
not string or int, as required by proto textformat.
It also no longer depends on the protobuf package.

The ProtoConfiguration data type was renamed from
"proto" to "proto_configuration" to avoid a conflict
error in the doc tools. (I don't know why the doc
tools have only a single namespace for data types
and functions---it's not as if int and bool are
surprising developments.)

RELNOTES:
Use proto.encode_text(x) to encode a Starlark value as textproto.
struct.to_proto() is deprecated and will be disabled by
the --incompatible_struct_has_no_methods flag.
Both functions now reject list/dict fields that contain list/dict elements.
PiperOrigin-RevId: 338518163